+++ src/Matrix.hs
@@ -71,16 +71,16 @@ normalise a = vectorMult a (sqrt (1/len a))
 
 -- Makes a Vertex3 (GLdouble) from a Vec3
 vert::Vec3->Vertex3(GLdouble)
-vert (x,y,z) = Vertex3 x y z
+vert (x,y,z) = Vertex3 (realToFrac x) (realToFrac y) (realToFrac z)
 
 -- Makes a TexCoord2 (GLdouble) from a Vec3
 coord::(Double,Double)->TexCoord2(GLdouble)
-coord (x,y) = TexCoord2 x y
+coord (x,y) = TexCoord2 (realToFrac x) (realToFrac y)
 
 -- Makes a Vector3 (GLdouble) from a Vec3
 vect::Vec3->Vector3(GLdouble)
-vect (x,y,z) = Vector3 x y z
+vect (x,y,z) = Vector3 (realToFrac x) (realToFrac y) (realToFrac z)
 
 -- Makes a Normal3 (GLdouble) from a Vec3
 norm3::Vec3->Normal3(GLdouble)
-norm3 (x,y,z) = Normal3 x y z
+norm3 (x,y,z) = Normal3 (realToFrac x) (realToFrac y) (realToFrac z)
